Determine voltage reading
I added some code to print out the voltage reading. It looks like the
voltage isn't quite what I expect it to be - I'm not sure whether this
is because the voltage reference isn't as stable as I want it to be, or
because there's noise, or something else. I'm hoping some of the noise
issues go away when I make a PCB (since I'll have decoupling caps
everywhere, and the wires won't all be acting as gigantic antennas).
I've also ordered voltage reference ICs, which can serve double duty as
the power supply for the instrumentation amplifier I ordered (since I
wasn't able to get my 3x op amp configuration working as expected,
probably due to needing to trim the resistors?)

There's also the calibration modes on the ADC to consider, which I
should probably look at once I have the basic setup working.

I also need to figure out if I can use a TRS cable, or perhaps TRRS
cable, to avoid needing to make a bunch of cables to send the voltage
output from the op amp back to the ADC. Initially, I was worried about
voltage drop / noise due to the distance (~3-4 ft per line), but I think
that this setup is pretty similar to mics work (mic level input which
gets preamp'd and then transmitted over a medium distance to the ADC),
so it should definitely be possible to do reasonably well. Maybe I'll go
find an audio engineer to talk to.

Cable-wise, I think I can get away with 5v and AGND input to the sensor
board (containing vref, op amp, and strain gauge), and then I'll need
either 1 or 2 lines for the output (depending on whether I want twisted
+ shielded output with a different ground line, to be used as a
differential input for better accuracy).